+++ CHANGES.md
@@ -31,6 +31,7 @@
 * New system features:
   * Super NES: Satellaview BS-X ROM headers are now decoded properly.
   * Wii: Unencrypted images from RVT-H Reader systems are now supported.
+  * NES: Improved publisher lookup for FDS titles.
 
 * New compressed texture formats:
   * Ericsson ETC1 and ETC2